Added initial physics classes.
The key class here is PhysicsProcessorBox2D which is a Box2D
implementation of the PhysicsProcessor interface. It is responsible for
initially creating all the physics bodies, then updating both them and
the GameState each iteration.
The PhysicsData class is the user data attached to each physics body.
There are agent and pill derived classes for additional user data
specific to bodies of these types.
PhysicsBodyType is just an enum used in the user data to distinguish
between different types of bodies.
PhysicsContact is the basic contact listener which flags bodies to
remove after contact.
PhysicsConfiguration contains the data that should be provided to
PhysicsProcessorBox2D on construction.