Initial groundwork for Genetic Algorithm
- The AI is currently a series of factors which are applied to
information about the current game state (direction to pills, prey and
predators, along with the distances to these bits of information). This
is encapsulated in the 'OrionAI' class. The class implements the AILogic
interface.
- Added the base GA classes (all the core classes, plus breeders,
mutators, and selectors), and their OrionAI implementations.
- Added basic infrastructure to log the results of games - the
GameResult and ResultLogger classes.
- Some changes to PredatorPreyGame to allow an external class to
manipulate its Pause/Running/etc. state, and also to reset the game.
This still needs some work to tidy it up.