Fix power ups for new movement mechanism
- The main change here is in PowerUpProcessor. When a power up is first
applied, or when it is about to end, we need to make sure the physics
body that the power up is applied to is in a viable position. The
position may be 'jogged' slightly to move it into a compatible position
if it is not already.
- Formalise use of debug categories for physics. You can now set a
physics debug enum in PredatorPreyGame which controls what debug
information is displayed.
- Removed the 'inTransition' status from Agents which isn't needed with
the new movement mechanism.
- Add some documentation.