Changes for Android
This commit makes changes which are required to get the game working on
android:
- Take out the JAXB xml parsing stuff. The default values in the
configuration file are used.
- Remove java.awt dependencies from PolygonShape class. (Some hacky
alternative algorithms are added for the time being - these should be
fixed properly).
- Replace Java 8 code from AILogicPartition with Andorid-supported code.
- Inject a 'CameraAccessor' into the input processor, so it can convert
screen coordinates to camera coords and access the current camera
position.  This allows us to work out where a 'tap' occurs on the
screen. The CameraAccessor interface is currently implemented by the
GameScreen.