Add basic sound infrastructure
This commit adds the basics of the infrastructure to add sound to the
game.
- A general callback system is added which allows classes to listen to
events broadcast by other classes (this is basically a stripped down
version of the Observer-Observable framework, but more tailored to our
needs).
- The PhysicsProcessor is made a 'Sender' and will broadcast 'Events' to
all 'Receiver' classes which have requested to be notified of physics
events.
- A SoundManager class is added which is a 'Receiver' of physics events.
When it hears about a physics event it chooses which sound to play
(which may be no sound).
- The visitor pattern is again used, this time for the PhysicsEvent
classes. See EventSoundProcessor.
- Currently there is just a PhysicsEventContact class which is sent to
receivers when a contact event occurs between any two bodies in the
world.
- I've added a dummy sound and some dummy music files to the assets
folder (all freeware .mp3s). These will be replaced later.