Re-work data management
This adds the first phase of work for the new data management system.
Note that, at this point, it's not complete and data is still read from
the xml config. But this puts in the first bit of work for swapping to a
different approach (the implementation will use a combination of .json
files and libGDX Preferences).