Hook up new data management system - Part 2
- This connects the sound data with the UI. Changing sound on/off and
volume levels on the UI now works.
- Basic music is added to the game. There are two placeholder music
types at the moment - one for the menu screens and one for the when the
game is played. The sound manager listens for changes of the UI screens
to check whether to change the music.
- Added maze generation parameters to the levels.
- For sandbox mode, the maze generation parameters are now randomised
between some reasonable values.
- Fixed a couple of problems with the game not working on Android. There
were capitalisation errors for a couple of image files in the
renderer_config.json file. Also, I'd used some Java 8 language features
which aren't supported on Android. This was some interfaces from
java.util.function. Since these are very simple interfaces, I've
replicated them in a new 'functional' package. If these features are
supported in Android in future, it should be easy to swap them in.