Fix bug setting Agent velocity when power up ends
When a Freeze power up terminated, if the (AI controlled) agent to which
it was applied had just entered a maze square, the agent may get stuck.
To solve this, if we can't work get a valid velocity for the agent by
other means, just move the agent to the nearest square centre and things
will get back on track.