Add level button stars
The buttons on the Levels screen now have three stars under each one to
indicate the player's progress.
This is primarily implemented by the new LevelButtonPanel class which
holds and positions the button and the three star Images.
The LevelsScreen gets the current player progress and the star scores
from the DataManager and updates the stars as appropriate (see
updateStars() method).
(This does not include implementing the Game score - a 'dummy' score is
currently set when a level is completed. This will be implemented
separately.)