Add score and time remaining
In the game screen, the score and time remaining now get updated. The
game ends if you run out of time.
Each level has a time limit, specified in the levels_config.json file.
The score is currently calculated based on the number of pills remaining
and the amount of time remaining.
The top score for a level gets saved out to the player progress when the
level is completed.