Initial commit for improved pathfinder
Firstly, all the pathfinding code is moved to its own package. This is
mostly independent of anything else in the game logic.
The concepts of a Path and a PathFinder are abstracted out into
interfaces. These are the only parts of the package that are exposed
publicly except for the the PathFinderCreator which is a factory which
allows external callers to request a particular type of PathFinder.
The AILogic code is updated to work with the interfaces.
A new concrete path finder is added, BFSPathFinder. This uses
breadth-first search to do an all-paths-single-source calculation for
each position in the maze, thereby generating all paths.
There is still work to be done to improve performance and tidy up code.