Initial work to add a Loading screen
This adds the concept of a ProgressTask. This is an interface which
extends Runnable. The implementing class should implement the run()
method where it carries out the task.  A ProgressTask can also be
interrogated for its current progress.
The PathFinder extends the ProgressTask interface. Derived classes
implement the run() method by calling generateAllPaths().
When the LoadingScreen is first shown, it asks the Game Logic for all
its ProgressTasks, then sets each one off on its own thread.
Each time the LoadingScreen is rendered, it asks each task for its
current progress and updates the Progress bar with the percent complete.