Tidy up assets
There are quite a few images which we don't use any more clogging up the
assets folder. This change removes the ones we don't use.
This also swaps the stars used on the levels screen to use the new star
images.
The old xml configuration file is removed as it's no longer used.