Add percent label and tidy up
- A label is added to the progress bar on the loading screen to show the
percent complete. This is wrapped up in the new ProgressBarPanel class.
- The PathFinder is no longer a ProgressTask. Instead a new
PathGenerator class is added which implements ProgressTask. It is
constructed with a PathFinder and calls generateAllPaths() when it's
asked to run().
- The PathFinder is changed from an interface to an abstract class which
holds the Maze being searched.
- The BFSPathFinder is changed to make sure the getPath() methods work
when generateAllPaths() hasn't been called. A bit of internal
refactoring is carried out to avoid duplicating code.