Initial work to revise main menu screen
This changes the layout of the main menu screen to have the game title
at the top, the buttons at the bottom and an AI-controlled 'preview'
game playing in the middle.
The CameraManager class currently used by the GameScreen has been
changed into the GameViewport class. This class now extends Viewport and
is used by both the GameScreen and the MainMenuScreen. The viewport can
be sized and positioned as required. The viewport's camera viewport and
camera position is set in the update() method.
The stages used by UI classes have been rationalised. The
AbstractScreen's stage is now only used for the screen background, this
stage uses a FillViewport. The MenuScreen now has a stage for all the UI
elements, which uses a ScreenViewport.
A new GameType has been added 'MainMenu' for the game playing on the
menu screen. The associated methods for getting the config classes has
been added to the GameDataManager.
The SoundManager doesn't play the game sounds for the main menu game.
It's now possible to set whether the renderer draws the maze background
or not.