Modify force game over approach
UserInputProcessor shouldn't really depend on PredatorPreyGame I don't
think. So I've removed the game member variable from this class. You can
still force the game to end via user input in the same way as before,
but now the UserInputProcessor stores the GameOver reason and adds it to
the current player Move, which is then examined by the game when it
processes the Move.