Disable rendering the game screen when game paused
This can cause problems for the win/lose scenarios. You end up in a
situation where the code is constantly trying to switch to the pause
screen. I'm disabling the rendering of the game in the background for
now - it will be fixed later after some upcoming changes to the screen
hierarchy.