Sort game viewport scaling
The GameViewport wasn't working correctly for different screen aspect
ratios. This class has now been changed so that it extends
ScalingViewport and uses the 'fill' scaling approach. This means all the
scaling is taken care for us. We now just have to make sure that the
camera position updates correctly. With this scaling strategy, the
viewport can extend beyond the screen edges, so there's logic to check
that the camera moves such that the maze stays on the screen and also
that the camera doesn't move beyond the maze.

The Main Menu now has its own game viewport class as it essentially
wants a different scaling strategy and camera management to the main
game viewport. It uses the 'fit' scaling strategy and allows limits to
be place on the y-extents of the screen in which the viewport can
position itself.