Remove CameraAccessor
The CameraAccessor interface was originally used to allow the
UserInputProcessor to access the game camera to convert between
world/screen coordinates and get the camera position. However, it's not
been in use for a while so I'm taking it out. It won't be hard to
re-introduce this or a similar mechanism if we need it again in future.