Change in-game dialogs
The in-game dialog design is changed such that instead of using a
separate PauseScreen class which tries to render the game in the
background and show a dialog on top, the GameScreen owns the dialogs and
just displays the right one at the right time. The PauseScreen has been
removed.

The dialogs have a base class, GameDialog, which provides methods for
some standard buttons (e.g. 'Restart' and 'Exit') to be added. It also
does the handling when a button is pressed. The derived classes then
just need to add the buttons they want in their constructors.

The game state now holds whether the game is currently running (i.e.
paused/finished/active), and the game over reason. The GameOver enum and
its states have been renamed for clarity to GameOverReason along with
PreyWon_Pills, PredatorWon, etc.

The GameStatus interface has been removed and the sound classes now just
store the PredatorPreyGame itself.  There's been some minor tidying of
the sound classes to make things a bit more consistent.